Responsive, Jsdoc, refresh token

+++ backend/src/model/user.ts
@@ -2,7 +2,16 @@ import { ROLE } from "../enums/role";
 import { GetUserQueryPage, User } from "../interfaces/user";
 import { BaseModel } from "./base";
 
+/**
+ * Class for handling user-related database operations.
+ */
 export class UserModel extends BaseModel {
+  /**
+   * Creates a new user in the database.
+   * @param user - The user object containing user details.
+   * @returns The result of the insert query.
+   * @throws Will throw an error if user creation fails.
+   */
   static async createUser(user: User) {
     const userToCreate = {
       name: user.name,
@@ -14,6 +23,12 @@ export class UserModel extends BaseModel {
     return query;
   }
 
+  /**
+   * Updates the password for a user identified by email.
+   * @param email - The email of the user whose password is to be updated.
+   * @param newPassword - The new password for the user.
+   * @throws Will throw an error if the update operation fails.
+   */
   static async updateUser(email: string, newPassword: string) {
     try {
       await this.queryBuilder()
@@ -26,19 +41,30 @@ export class UserModel extends BaseModel {
     }
   }
 
+  /**
+   * Retrieves a list of users based on the filter criteria.
+   * @param filter - An object containing filter criteria, pagination, and search query.
+   * @returns A query object for the list of users.
+   */
   static getUsers(filter: GetUserQueryPage) {
     const { q } = filter;
+    console.log(filter.page);
     const query = this.queryBuilder()
       .select("userId", "name", "email")
       .table("users")
-      .limit(filter.size)
-      .offset((filter.page - 1) * filter.size);
+      .limit(filter.limit)
+      .offset((filter.page - 1) * filter.limit);
     if (q) {
       query.whereLike("name", `%${q}%`);
     }
     return query;
   }
 
+  /**
+   * Counts the total number of users based on the filter criteria.
+   * @param filter - An object containing filter criteria and search query.
+   * @returns A query object for the user count.
+   */
   static count(filter: GetUserQueryPage) {
     const { q } = filter;
     const query = this.queryBuilder().count("*").table("users").first();
@@ -48,6 +74,11 @@ export class UserModel extends BaseModel {
     return query;
   }
 
+  /**
+   * Retrieves a user by their user ID.
+   * @param userId - The ID of the user to retrieve.
+   * @returns A query object for the user details.
+   */
   static getUserById(userId: string) {
     console.log("userid" + userId);
     const query = this.queryBuilder()
@@ -58,6 +89,11 @@ export class UserModel extends BaseModel {
     return query;
   }
 
+  /**
+   * Retrieves a user by their email.
+   * @param email - The email of the user to retrieve.
+   * @returns A query object containing user details or null if no user is found.
+   */
   static async getUserByEmail(email: string) {
     const query = await this.queryBuilder()
       .select("userId", "name", "email", "password", "role")
@@ -67,7 +103,22 @@ export class UserModel extends BaseModel {
     return query;
   }
 
+  /**
+   * Deletes a user by their user ID.
+   * @param userId - The ID of the user to delete.
+   * @returns The result of the delete query.
+   * @throws Will throw an error if trying to delete an admin user.
+   */
   static async deleteUser(userId: string) {
+    const userToDelete = await this.queryBuilder()
+      .select("role")
+      .table("users")
+      .where({ userId })
+      .first();
+
+    if (userToDelete && userToDelete.role === ROLE.ADMIN) {
+      throw new Error("Cannot delete an admin user");
+    }
     const query = await this.queryBuilder()
       .delete()
       .table("users")
